+++ bin/oprah
@@ -5,7 +5,7 @@
 const chalk = require('chalk');
 const pjson = require('../package.json');
 const program = require('commander');
-const { makeOprah } = require('../lib/makeOprah');
+const { makeOprah } = require('../lib/make-oprah');
 const { logError, log } = require('../lib/utils/logger');
 
 let oprahPromise = null;
