+++ lib/commands/init/make-init.test.js
@@ -6,7 +6,7 @@ jest.mock('../../services/cf/deploy', () => ({
   deploy: mockDeploy
 }));
 
-const { makeOprah } = require('../../makeOprah');
+const { makeOprah } = require('../../make-oprah');
 
 describe('init', () => {
   beforeEach(() => {
