+++ lib/commands/list/make-list.test.js
@@ -16,7 +16,7 @@ jest.mock('./list-parameters', () => ({
 
 const path = require('path');
 
-const { makeOprah } = require('../../makeOprah');
+const { makeOprah } = require('../../make-oprah');
 
 describe('list', () => {
   let promise;
