+++ lib/commands/run/make-run.test.js
@@ -25,7 +25,7 @@ jest.mock('../../services/parameter-store/make-parameter-store', () => ({
 
 const path = require('path');
 
-const { makeOprah } = require('../../makeOprah');
+const { makeOprah } = require('../../make-oprah');
 
 describe('run', () => {
   let promise;
